Split resource monitors into dedicated files
